+++ src/main/java/org/ecommerce/repositories/CategoryRepository.java
@@ -1,7 +1,7 @@
 package org.ecommerce.repositories;
 
 import org.ecommerce.models.Category;
-import org.ecommerce.services.impl.CrudOperationsImpl;
+import org.ecommerce.repositories.impl.CrudOperationsImpl;
 
 
 public class CategoryRepository extends CrudOperationsImpl<Category> {
