+++ src/main/java/org/ecommerce/repositories/OrderRepository.java
@@ -1,7 +1,7 @@
 package org.ecommerce.repositories;
 
 import org.ecommerce.models.Order;
-import org.ecommerce.services.impl.CrudOperationsImpl;
+import org.ecommerce.repositories.impl.CrudOperationsImpl;
 
 public class OrderRepository extends CrudOperationsImpl<Order> {
