+++ src/main/java/org/ecommerce/repositories/ProductRepository.java
@@ -1,7 +1,7 @@
 package org.ecommerce.repositories;
 
 import org.ecommerce.models.Product;
-import org.ecommerce.services.impl.CrudOperationsImpl;
+import org.ecommerce.repositories.impl.CrudOperationsImpl;
 
 public class ProductRepository extends CrudOperationsImpl<Product> {
 }
