+++ src/main/java/org/ecommerce/services/UserService.java
@@ -2,6 +2,6 @@ package org.ecommerce.services;
 
 import org.ecommerce.models.User;
 
-public interface UserService<T extends User> extends CrudOperations<T, Long> {
+public interface UserService<T extends User> extends OperationsService<T, Long> {
 
 }
